+++ noodles-sam/src/header/reference_sequence.rs
@@ -305,13 +305,13 @@ impl ReferenceSequence {
     /// let reference_sequence = ReferenceSequence::builder()
     ///     .set_name("sq0")
     ///     .set_length(13)
-    ///     .insert(Tag::Other(String::from("zn")), String::from("noodles"))
+    ///     .insert(Tag::Other([b'z', b'n']), String::from("noodles"))
     ///     .build()?;
     ///
     /// let fields = reference_sequence.fields();
     /// assert_eq!(fields.len(), 1);
     /// assert_eq!(
-    ///     fields.get(&Tag::Other(String::from("zn"))),
+    ///     fields.get(&Tag::Other([b'z', b'n'])),
     ///     Some(&String::from("noodles"))
     /// );
     ///