+++ noodles-sam/src/header/reference_sequence/tag.rs
@@ -1,6 +1,13 @@
 //! SAM header reference sequence tag.
 
-use std::{error, fmt, str::FromStr};
+use std::{
+    convert::TryFrom,
+    error,
+    fmt::{self, Write},
+    str::FromStr,
+};
+
+const LENGTH: usize = 2;
 
 /// A SAM header reference sequence tag.
 #[derive(Clone, Debug, Hash, Eq, PartialEq)]
@@ -26,30 +33,33 @@ pub enum Tag {
     /// URI of the reference sequence (`UR`).
     Uri,
     /// Any other reference sequence tag.
-    Other(String),
+    Other([u8; 2]),
 }
 
-impl AsRef<str> for Tag {
-    fn as_ref(&self) -> &str {
+impl AsRef<[u8; LENGTH]> for Tag {
+    fn as_ref(&self) -> &[u8; LENGTH] {
         match self {
-            Self::Name => "SN",
-            Self::Length => "LN",
-            Self::AlternativeLocus => "AH",
-            Self::AlternativeNames => "AN",
-            Self::AssemblyId => "AS",
-            Self::Description => "DS",
-            Self::Md5Checksum => "M5",
-            Self::Species => "SP",
-            Self::MoleculeTopology => "TP",
-            Self::Uri => "UR",
-            Self::Other(s) => s,
+            Self::Name => b"SN",
+            Self::Length => b"LN",
+            Self::AlternativeLocus => b"AH",
+            Self::AlternativeNames => b"AN",
+            Self::AssemblyId => b"AS",
+            Self::Description => b"DS",
+            Self::Md5Checksum => b"M5",
+            Self::Species => b"SP",
+            Self::MoleculeTopology => b"TP",
+            Self::Uri => b"UR",
+            Self::Other(tag) => tag,
         }
     }
 }
 
 impl fmt::Display for Tag {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.as_ref())
+        let bytes = self.as_ref();
+        f.write_char(char::from(bytes[0]))?;
+        f.write_char(char::from(bytes[1]))?;
+        Ok(())
     }
 }
 
@@ -77,25 +87,32 @@ impl FromStr for Tag {
     type Err = ParseError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "" => Err(ParseError::Empty),
-            "SN" => Ok(Self::Name),
-            "LN" => Ok(Self::Length),
-            "AH" => Ok(Self::AlternativeLocus),
-            "AN" => Ok(Self::AlternativeNames),
-            "AS" => Ok(Self::AssemblyId),
-            "DS" => Ok(Self::Description),
-            "M5" => Ok(Self::Md5Checksum),
-            "SP" => Ok(Self::Species),
-            "TP" => Ok(Self::MoleculeTopology),
-            "UR" => Ok(Self::Uri),
-            _ => {
-                if s.len() == 2 {
-                    Ok(Self::Other(s.into()))
-                } else {
-                    Err(ParseError::Invalid)
-                }
-            }
+        let bytes = s.as_bytes();
+
+        match bytes.len() {
+            0 => Err(ParseError::Empty),
+            2 => Self::try_from([bytes[0], bytes[1]]),
+            _ => Err(ParseError::Invalid),
+        }
+    }
+}
+
+impl TryFrom<[u8; LENGTH]> for Tag {
+    type Error = ParseError;
+
+    fn try_from(b: [u8; LENGTH]) -> Result<Self, Self::Error> {
+        match &b {
+            b"SN" => Ok(Self::Name),
+            b"LN" => Ok(Self::Length),
+            b"AH" => Ok(Self::AlternativeLocus),
+            b"AN" => Ok(Self::AlternativeNames),
+            b"AS" => Ok(Self::AssemblyId),
+            b"DS" => Ok(Self::Description),
+            b"M5" => Ok(Self::Md5Checksum),
+            b"SP" => Ok(Self::Species),
+            b"TP" => Ok(Self::MoleculeTopology),
+            b"UR" => Ok(Self::Uri),
+            _ => Ok(Self::Other(b)),
         }
     }
 }
@@ -116,7 +133,7 @@ mod tests {
         assert_eq!(Tag::Species.to_string(), "SP");
         assert_eq!(Tag::MoleculeTopology.to_string(), "TP");
         assert_eq!(Tag::Uri.to_string(), "UR");
-        assert_eq!(Tag::Other(String::from("ND")).to_string(), "ND");
+        assert_eq!(Tag::Other([b'N', b'D']).to_string(), "ND");
     }
 
     #[test]
@@ -131,7 +148,7 @@ mod tests {
         assert_eq!("SP".parse(), Ok(Tag::Species));
         assert_eq!("TP".parse(), Ok(Tag::MoleculeTopology));
         assert_eq!("UR".parse(), Ok(Tag::Uri));
-        assert_eq!("ND".parse(), Ok(Tag::Other(String::from("ND"))));
+        assert_eq!("ND".parse(), Ok(Tag::Other([b'N', b'D'])));
 
         assert_eq!("".parse::<Tag>(), Err(ParseError::Empty));
         assert_eq!("NDL".parse::<Tag>(), Err(ParseError::Invalid));